cardscreen 랜덤으로 컨텐츠 지정
이전 스크롤이 지속되지 않는다는 문제점
