+++ src/assets/dummydata.js
@@ -5,6 +5,8 @@ export const VIDEO_DATA = [
     url: 'https://www.w3schools.com/html/mov_bbb.mp4',
     description: 'This is a sample video description.',
     thumbnail: 'https://via.placeholder.com/150',
+    date: '2024-06-01',
+    source: '한국일보 외 2',
   },
   {
     id: 2,
@@ -12,6 +14,8 @@ export const VIDEO_DATA = [
     url: 'https://www.w3schools.com/html/movie.mp4',
     description: 'Another sample video.',
     thumbnail: 'https://via.placeholder.com/150',
+    date: '2024-06-02',
+    source: '123123일보 외 2',
   },
   {
     id: 3,
@@ -19,5 +23,7 @@ export const VIDEO_DATA = [
     url: 'https://www.w3schools.com/html/mov_bbb.mp4',
     description: 'Third sample video.',
     thumbnail: 'https://via.placeholder.com/150',
+    date: '2024-06-03',
+    source: '123일보 외 2',
   },
 ];
